refactor: move mock authentication to helper
